refactored opportunityRoutes in api routes folder under controllers
