+++ controllers/homeRoutes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { Opportunity, User } = require('./models');
+const { Opportunity, User } = require('../models');
 const withAuth = require('./utils/auth');
 
 router.get('/', async (req, res) => {
